Bad column type

+++ scripts/pgsql.sql
@@ -83,7 +83,7 @@ DROP TABLE IF EXISTS galette_events_activitiesevents CASCADE;
 CREATE TABLE galette_events_activitiesevents (
   id_event integer NOT NULL,
   id_activity integer NOT NULL,
-  status tinyint(1) NOT NULL,
+  status smallint NOT NULL,
   PRIMARY KEY(id_event,id_activity)
 );
